move remote redirect tests to HttpClientHandler_RemoteServerTest
